Bugfix: Moved countable_resource out of utility.
Moved countable_resource out of utility as it requires atomic and that
is not compatible with wasi
